+++ .github/workflows/scripts/linux_arm64/rh8/build_catalyst.sh
@@ -46,7 +46,7 @@ cmake -S runtime -B runtime-build -G Ninja \
     -DPYTHON_LIBRARY=/opt/_internal/cpython-${PYTHON_VERSION}.${PYTHON_SUBVERSION}/lib \
     -Dpybind11_DIR=/opt/_internal/cpython-${PYTHON_VERSION}.${PYTHON_SUBVERSION}/lib/python${PYTHON_VERSION}/site-packages/pybind11/share/cmake/pybind11 \
     -DENABLE_LIGHTNING_KOKKOS=ON \
-    -DLIGHTNING_GIT_TAG=latest_release \
+    -DLIGHTNING_GIT_TAG=cfc3006415c695aff5237ffbe82510d1d608f781 \
     -DENABLE_LAPACK=OFF \
     -DKokkos_ENABLE_SERIAL=ON \
     -DKokkos_ENABLE_OPENMP=ON \