+++ runtime/lib/backend/lightning/lightning_kokkos/LightningKokkosSimulator.cpp
@@ -126,6 +126,19 @@ void LightningKokkosSimulator::PrintState()
     cout << state[idx] << "]" << endl;
 }
 
+void LightningKokkosSimulator::SetState(DataView<std::complex<double>, 1> &data_view)
+{
+    std::vector<Kokkos::complex<double>> data_vector(data_view.begin(), data_view.end());
+    std::vector<std::size_t> count(data_vector.size());
+    std::iota(std::begin(count), std::end(count), 0);
+    this->device_sv->setStateVector(count, data_vector);
+}
+
+void LightningKokkosSimulator::SetBasisState(const std::size_t index)
+{
+    this->device_sv->setBasisState(index);
+}
+
 auto LightningKokkosSimulator::Zero() const -> Result
 {
     return const_cast<Result>(&GLOBAL_RESULT_FALSE_CONST);