+++ runtime/lib/backend/openqasm/OpenQasmDevice.cpp
@@ -116,6 +116,9 @@ void OpenQasmDevice::PrintState()
     cout << state[idx] << "]" << endl;
 }
 
+void OpenQasmDevice::SetState(DataView<std::complex<double>, 1> &) {}
+void OpenQasmDevice::SetBasisState(const std::size_t index) {}
+
 auto OpenQasmDevice::Zero() const -> Result
 {
     return const_cast<Result>(&GLOBAL_RESULT_FALSE_CONST);