+++ doc/changelog.md
@@ -186,6 +186,11 @@
 
 * Update JAX to `v0.4.28`. [(#931)](https://github.com/PennyLaneAI/catalyst/pull/931)
 
+* Allows for simulators to specify the optimization of setting the initial state directly
+  via `qml.StatePrep` and `qml.BasisState` when using the `initial_state_prep` flag in the
+  TOML file.
+  [(#955)](https://github.com/PennyLaneAI/catalyst/pull/955)
+
 <h3>Breaking changes</h3>
 
 * Return values of qjit-compiled functions that were previously `numpy.ndarray` are now of type
@@ -201,6 +206,10 @@
   function is valid. Keyword arguments can be passed to this function using the
   `extrapolate_kwargs` keyword argument in `mitigate_with_zne`.
   [(#806)](https://github.com/PennyLaneAI/catalyst/pull/806)
+
+* The QuantumDevice API has now added the functions `SetState` and `SetBasisState`
+  for simulators that may benefit from instructions that directly set the state.
+  [(#955)](https://github.com/PennyLaneAI/catalyst/pull/955)
 
 <h3>Bug fixes</h3>